client/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.9
Pillow==8.4.0
psycopg==3.0.1
psycopg2-binary==2.8.6
pytz==2021.3
sqlparse==0.4.2
djangorestframework
requests